Changed some types

+++ src/InputDevice/InputDevice.cpp
@@ -1,7 +1,8 @@
+#include <cstdint>
 #include <initializer_list>
 #include "InputDevice.hpp"
 
 namespace Intel8080 {
-    InputDevice::InputDevice(const std::initializer_list<int>& inputPortNumbers)
+    InputDevice::InputDevice(const std::initializer_list<uint8_t>& inputPortNumbers)
         : inputPortNumbers{inputPortNumbers}{}
 }
