+++ src/OutputDevice/OutputDevice.cpp
@@ -2,6 +2,6 @@
 #include "OutputDevice.hpp"
 
 namespace Intel8080 {
-    OutputDevice::OutputDevice(const std::initializer_list<int>& outputPortNumbers)
+    OutputDevice::OutputDevice(const std::initializer_list<uint8_t>& outputPortNumbers)
         : outputPortNumbers{outputPortNumbers}{}
 }
