+++ src/OutputDevice/OutputDevice.hpp
@@ -8,12 +8,12 @@
 namespace Intel8080 {
     class OutputDevice {
         public:
-            OutputDevice(const std::initializer_list<int>& outputPortNumbers);
+            OutputDevice(const std::initializer_list<uint8_t>& outputPortNumbers);
             virtual ~OutputDevice(){}
 
-            const std::vector<int> outputPortNumbers;
+            const std::vector<uint8_t> outputPortNumbers;
 
-            virtual void writeByte(int portNumber, uint8_t byte) = 0;
+            virtual void writeByte(uint8_t portNumber, uint8_t byte) = 0;
     };
 }
